requirements.txt
-i https://pypi.org/simple
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.0.4
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
pyjwt==1.7.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
sqlalchemy==1.3.19
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1
wtforms==2.3.3
